warning fixes

+++ test/test_file_io.hpp
@@ -35,7 +35,7 @@ struct TestInput : public quicker_sfv::FileInput {
     }
 
     int64_t seek(int64_t offset, SeekStart seek_start) override {
-        int64_t base_index = 0;
+        size_t base_index = 0;
         if (seek_start == SeekStart::CurrentPosition) {
             base_index = read_idx;
         } else if (seek_start == SeekStart::FileEnd) {
@@ -45,7 +45,7 @@ struct TestInput : public quicker_sfv::FileInput {
         if ((new_index < 0) || (new_index > static_cast<int64_t>(contents.size()))) {
             return -1;
         }
-        read_idx = new_index;
+        read_idx = static_cast<size_t>(new_index);
         return new_index;
     }
